+++ apps/web/src/app/api/admin/workers/[queueName]/route.ts
@@ -5,10 +5,7 @@ import { NextRequest, NextResponse } from 'next/server'
 
 export const GET = errorHandler(
   adminHandler(
-    async (
-      _: NextRequest,
-      { params }: { params: { queueName: string } },
-    ) => {
+    async (_: NextRequest, { params }: { params: { queueName: string } }) => {
       const result = await getQueueDetail({ queueName: params.queueName })
       const detail = result.unwrap()
       return NextResponse.json(detail, { status: 200 })