+++ packages/telemetry/typescript/src/sdk/index.ts
@@ -1,2 +1,17 @@
+export type {
+  CaptureOptions,
+  ChatSpanOptions,
+  EndCompletionSpanOptions,
+  EndHttpSpanOptions,
+  EndSpanOptions,
+  EndToolSpanOptions,
+  ErrorOptions,
+  ExternalSpanOptions,
+  PromptSpanOptions as PromptSegmentOptions,
+  StartCompletionSpanOptions,
+  StartHttpSpanOptions,
+  StartSpanOptions,
+  StartToolSpanOptions,
+} from '$telemetry/instrumentations'
 export * from './redact'
 export * from './sdk'